+++ CHANGELOG.md
@@ -7,7 +7,8 @@
    - Add isSupportColor, setPrintOptions, reset (allow reusing the same print task for multiple print jobs)
    - Add options.cutHeight
    - Fix options.cutType not being used
-* Add printheadWidth gathering on printer connection (PrinterInfo object)
+* Add printheadWidth, supportColor gathering on printer connection (PrinterInfo object)
+* Add double color support to B1PrintTask
 
 # 0.0.1-alpha.44